Refactor the code to make it more modularized
