+++ .specify/memory/constitution.md
@@ -1,0 +1,37 @@
+<!--
+Sync Impact Report:
+- Version change: 1.0.0
+- Added Principles:
+  - I. Performant and Clean Code
+  - II. Material Design
+  - III. Pragmatic Testing
+- Templates requiring updates:
+  - ✅ /Users/milan/Desktop/ Learn/Services_App/.specify/templates/plan-template.md
+  - ✅ /Users/milan/Desktop/ Learn/Services_App/.specify/templates/tasks-template.md
+  - ✅ /Users/milan/Desktop/ Learn/Services_App/README.md
+-->
+# Services_App Constitution
+
+## Core Principles
+
+### I. Performant and Clean Code
+All code must be written with performance and readability in mind. This includes following Flutter best practices, avoiding unnecessary computations, and ensuring the codebase is easy to maintain and understand.
+
+### II. Material Design
+The application's user interface will adhere to Google's Material Design guidelines. Use and customize the MaterialApp theme to ensure a consistent and high-quality user experience.
+
+### III. Pragmatic Testing
+While Test-Driven Development (TDD) is not a strict requirement, writing tests is highly encouraged. Focus on creating unit and widget tests for critical parts of the application to ensure stability and prevent regressions.
+
+## Additional Constraints
+<!-- This section is a placeholder for any additional project-wide constraints. -->
+[SECTION_2_CONTENT]
+
+## Development Workflow
+<!-- This section is a placeholder for the project's development workflow. -->
+[SECTION_3_CONTENT]
+
+## Governance
+All pull requests and reviews must verify compliance with the constitution. Complexity must be justified.
+
+**Version**: 1.0.0 | **Ratified**: 2025-09-30 | **Last Amended**: 2025-09-30
