+++ README.md
@@ -1,4 +1,4 @@
-# locker_app
+# Services_App
 
 A new Flutter project.
