+++ lib/src/features/safe_home/widgets/file_viewers.dart
@@ -1,0 +1,101 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:just_audio/just_audio.dart';
+import 'package:video_player/video_player.dart';
+
+class ImageViewer extends StatelessWidget {
+  const ImageViewer({super.key, required this.file});
+  final File file;
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(file.uri.pathSegments.last)),
+      body: Center(child: Image.file(file, fit: BoxFit.contain)),
+    );
+  }
+}
+
+class VideoViewer extends StatefulWidget {
+  const VideoViewer({super.key, required this.file});
+  final File file;
+  @override
+  State<VideoViewer> createState() => _VideoViewerState();
+}
+
+class _VideoViewerState extends State<VideoViewer> {
+  late final VideoPlayerController _controller;
+  @override
+  void initState() {
+    super.initState();
+    _controller = VideoPlayerController.file(widget.file)
+      ..initialize().then((_) {
+        setState(() {});
+        _controller.play();
+      });
+  }
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(widget.file.uri.pathSegments.last)),
+      body: Center(
+        child: _controller.value.isInitialized
+            ? AspectRatio(aspectRatio: _controller.value.aspectRatio, child: VideoPlayer(_controller))
+            : const CircularProgressIndicator(),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          setState(() {
+            _controller.value.isPlaying ? _controller.pause() : _controller.play();
+          });
+        },
+        child: Icon(_controller.value.isPlaying ? Icons.pause : Icons.play_arrow),
+      ),
+    );
+  }
+}
+
+class AudioPlayerScreen extends StatefulWidget {
+  const AudioPlayerScreen({super.key, required this.file});
+  final File file;
+  @override
+  State<AudioPlayerScreen> createState() => _AudioPlayerScreenState();
+}
+
+class _AudioPlayerScreenState extends State<AudioPlayerScreen> {
+  final AudioPlayer _player = AudioPlayer();
+  @override
+  void initState() {
+    super.initState();
+    _player.setFilePath(widget.file.path);
+  }
+  @override
+  void dispose() {
+    _player.dispose();
+    super.dispose();
+  }
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(widget.file.uri.pathSegments.last)),
+      body: Center(
+        child: StreamBuilder<PlayerState>(
+          stream: _player.playerStateStream,
+          builder: (context, snap) {
+            final playing = snap.data?.playing ?? false;
+            return IconButton(
+              iconSize: 64,
+              icon: Icon(playing ? Icons.pause_circle : Icons.play_circle),
+              onPressed: () => playing ? _player.pause() : _player.play(),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
