+++ specs/001-code-refactoring-i/spec.md
@@ -1,0 +1,35 @@
+# Feature Specification: Code Refactoring
+
+**Feature Branch**: `001-code-refactoring-i`
+**Created**: 2025-09-30
+**Status**: Draft
+**Input**: User description: "Code refactoring. I want to refactor the codebase. Only main.dart file exists now. I need the code to be more cleaner. The current code is a bit redundant. Refactor it with best practices."
+
+## User Scenarios & Testing *(mandatory)*
+
+### Primary User Story
+As a developer, I want to refactor the `main.dart` file to improve code quality, reduce redundancy, and adhere to Flutter best practices, so that the codebase is easier to maintain and understand.
+
+### Acceptance Scenarios
+1. **Given** the current `main.dart` file, **When** the refactoring is complete, **Then** the code should be organized into smaller, more manageable widgets and classes.
+2. **Given** the refactored code, **When** the app is run, **Then** all existing functionality should work as before.
+3. **Given** the refactored code, **When** analyzed, **Then** it should demonstrate improved readability and adherence to the project's constitution (Performant and Clean Code, Material Design).
+
+### Edge Cases
+- The refactoring should not introduce any new bugs or regressions.
+- The app's performance should not be negatively impacted by the refactoring.
+
+## Requirements *(mandatory)*
+
+### Functional Requirements
+- **FR-001**: The refactored code MUST maintain all existing functionality of the application.
+- **FR-002**: The code MUST be reorganized into a more logical structure, separating UI, business logic, and data management.
+- **FR-003**: Redundant code SHOULD be eliminated.
+- **FR-004**: The code MUST follow Flutter best practices and the project's coding standards.
+
+### Key Entities *(include if feature involves data)*
+- **ServicesApp**: The main application widget.
+- **LockScreen**: The screen for unlocking the application.
+- **SafeHome**: The main screen after unlocking, displaying files.
+- **SettingsStore**: A class for managing application settings.
+- **File Viewers**: Widgets for viewing images, videos, and audio files.
